Added _New (Blank) board to the ChooseGame list
Added the _New (Blank) board to the ChooseGame list so user can choose that one.  Added restrictions to prevent overwriting this board or naming something too similar to it.  Added file name save validation:
- Character count 3-30
- No special characters
- Certain filenames restricted

The PassSaveValidation was removed from Update as two validations were taking place and only need to validate when the user clicks on Save Board.