More details for choose game buttons
Provides more details around the various game boards that are available for play such as tile count.

Does not allow the button for the _New (Blank) board to be loaded if the user has selected "Play" from the main menu.  This ensures that only a playable board can be loaded.