Number of Players can be selected when Play is selected from Main Menu
The number of players on the ChooseGame scene can be selected when Play is selected from the main menu.  The user is then dropped into a game with those players.  Did the offsets of players 5-8 by just scaling the offset for players 1-4 so players 5-8 will appear to be more in the middle of the tile.

Added a simple camera system so much much larger boards can be built and played.  W A S D are used for panning up, left, down, and right.  The mouse scroll wheel also works.

TODO:
In the Start method of Play or CreateEdit scenes we need to do some math to determine the proper camera setting to show the whole board which is currently a problem with bigger boards.  Also, need to add some script to make the Camera Manager go to the position of the player whose turn it is.  Then the camera will move and center on the current player which would be better, especially for really big boards.