Cleaned up the Play scene a bit
Removed extraneous details that were displayed such as the start and finish tiles as well as the tile count.  Added a panel behind the UI where the player's turn, roll result, and Roll D4/Roll D6 buttons are shown so they could be more easily distinguished from the board.  Changed the color of the who wins text to make it stand out more.