Attempt at Online Multiplayer
Attempted to implement online multiplayer.  I feel like I am close, but need to implement some interface code into various  classes so the Photon Engine updates the player sprite and position.  New players are instantiated properly except for their offset and the token.  I think that's because during instantiation, Photon picks up the change, but immediately after instantiation I'm updating the token with the proper player's token as well as the player number.  Those changes after the instantiation aren't being picked up and one token is not being updated in the scene of the other game instance.  On the main menu, I've deactivated the button for "Online Multiplayer" called Loading.  Feel free to experiment by activating that button and playing/making further changes.  I'm committing this build because the game can be built and it looks like for Unity projects, only standalone files (.exe) are allowed.